Cleans up project.py.
